requirements.txt
Flask==1.1.1
cv2==4.1.1
tensorflow==2.0.0
ffmpeg-python==0.2.0


pip install ffmpeg-python
